// projects/student/src/app/api/main/v1/model/user.ts
/**
 * UG Examine
 * UG Examine API description
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface User { 
    username: string;
    password: string;
    fullName?: string;
    email: string;
    phoneNumber: string;
    role: User.RoleEnum;
}
export namespace User {
    export type RoleEnum = 'student' | 'staff' | 'admin';
    export const RoleEnum = {
        Student: 'student' as RoleEnum,
        Staff: 'staff' as RoleEnum,
        Admin: 'admin' as RoleEnum
    };
}